+++ semana18/projeto-labECommerce/.vscode/settings.json
@@ -1,8 +1,12 @@
 {
     "cSpell.words": [
+        "conex",
+        "conexão",
         "dotenv",
+        "estabelece",
         "knex",
         "labecommerce",
+        "o",
         "projeto"
     ]
 }
