+++ semana18/projeto-labECommerce/connection/connection.ts
@@ -1,0 +1,20 @@
+
+import knex from "knex";
+import dotenv from "dotenv";
+
+dotenv.config()
+
+
+// Estabelece conexão com o banco
+const connection = knex({
+    client: "mysql",
+    connection: {
+        host: process.env.DB_HOST,
+        port: 3306,
+        user: process.env.DB_USER,
+        password: process.env.DB_PASSWORD,
+        database: process.env.DB_SCHEMA,
+        multipleStatements: true
+    }
+})
+export default connection
